Minify macros by introducing store-level operation macro
